Enum menhir::MenhirOption

pub enum MenhirOption {
    Base(PathBuf),
    Canonical,
    CompileErrors(PathBuf),
    Dump,
    Explain,
    Graph,
    Lalr,
    LogAutomaton(usize),
    LogCode(usize),
    LogGrammar(usize),
    NoInline,
    NoStdlib,
    OnlyPreprocess(OnlyPreprocessOption),
    Stdlib(PathBuf),
    Strict,
    Suggest(SuggestOption),
    Timings,
    UnusedToken(String),
    UnusedTokens,
    Version,
    Verbose,
}

An option (flag) to be passed to the Menhir generator

Variants
Base(PathBuf)

The base path (without the .rs extension) of the output file

Canonical

Construct a canonical Knuth LR(1) automaton

CompileErrors(PathBuf)

Compile a .messages error file to Rust code

Dump

Describe the automaton in <basename>.automaton

Explain

Explain conflicts in <basename>.conflicts

Graph

Write grammar's dependency graph to <basename>.dot

Lalr

Construct an LALR(1) automaton

LogAutomaton(usize)

Log information about the automaton

LogCode(usize)

Log information about the generated code

LogGrammar(usize)

Log information about the grammar

NoInline

Ignore the %inline keyword

NoStdlib

Do not load the standard library

This is currently the default for the Rust backend when using process_file and compile_errors since the standard library has not been ported to Rust yet. You may want to use Stdlib if you want to load your own standard library.

OnlyPreprocess(OnlyPreprocessOption)

Only process the given stages and exit ; see OnlyPreprocessOption

Stdlib(PathBuf)

Specify where the standard library lies

Strict

Warnings about the grammar are errors

Suggest(SuggestOption)

Suggest where various files are located ; see SuggestOption

Timings

Display internal timings

UnusedToken(String)

Do not warn that the given token is unused

UnusedTokens

Do not warn about any unused token

Version

Show version number and exit

Verbose

Synonymous with Dump and Explain