Function menhir::compile_errors

pub fn compile_errors(file: &Path, grammar: &Path, args: &[MenhirOption])

Convenience function over run to compiles error files

Generate an output file in the OUT_DIR directory of the package that is currently being built. This function should be invoked from a Cargo build.rs build script.